requirements.txt
pandas
openpyxl
reportlab
pillow
requests
opencv-python
pillow_heif
emoji
pypdf
